requirements.txt
Django>=4.2.6,<4.3
djangorestframework>=3.14,<3.15
django-filter>=23.3,<23.4
markdown>=3.5,<3.6
psycopg2
drf-spectacular
djangorestframework-simplejwt